feat: Implement role-based access control with super_admin role
- Add super_admin role to Worker type
- Add requireSuperAdmin() auth function
- Update worker management APIs to require super_admin
- Fix decimal quantity parsing (parseInt to parseFloat)
- Create migration endpoint for promoting admin to super_admin

+++ components/worker/worker-dashboard.tsx
@@ -86,7 +86,7 @@ export function WorkerDashboard({ worker }: WorkerDashboardProps) {
         headers: { "Content-Type": "application/json" },
         body: JSON.stringify({
           product_name: productName,
-          quantity: Number.parseInt(quantity),
+          quantity: Number.parseFloat(quantity),
           unit_price: Number.parseFloat(unitPrice),
           notes: notes || null,
           unit_type: unitType,
